can login

+++ app/models/User.scala
@@ -1,4 +1,4 @@
 package models
 
-case class User (id: Long, login: String, password_hash: String, salt: String)
+case class User(id: Int, login: String, password_hash: String, salt: String)
 case class DisplayUser(login: String, ip: String, createdAt: String)
